+++ accounts/models.py
@@ -1,0 +1,23 @@
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+
+class Position(models.Model):
+    name = models.CharField(max_length=100, unique=True)
+
+    def __str__(self):
+        return self.name
+
+
+class Worker(AbstractUser):
+    # Custom user model with optional position
+    position = models.ForeignKey(
+        Position,
+        on_delete=models.SET_NULL,
+        null=True,
+        blank=True,
+        related_name="workers"
+    )
+
+    def __str__(self):
+        return self.username
