+++ config/urls.py
@@ -18,11 +18,10 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 
-from manager.views import SignUpView
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("", include("manager.urls")),
+    path("accounts/", include("accounts.urls")),
     path("accounts/", include("django.contrib.auth.urls")),
-    path("accounts/signup/", SignUpView.as_view(), name="signup"),
 ]
