+++ manager/admin.py
@@ -1,36 +1,19 @@
 from django.contrib import admin
-from .models import Position, Worker, TaskType, Task, Tag, Comment, Project
-
-# Admin config for Position model
-@admin.register(Position)
-class PositionAdmin(admin.ModelAdmin):
-    list_display = ("id", "name")
-    search_fields = ("name",)
+from .models import TaskType, Task, Tag, Comment, Project
 
 
-# Admin config for custom User model (Worker)
-@admin.register(Worker)
-class WorkerAdmin(admin.ModelAdmin):
-    list_display = ("id", "username", "email", "position")
-    search_fields = ("username", "email")
-    list_filter = ("position",)
-
-
-# Admin config for TaskType
 @admin.register(TaskType)
 class TaskTypeAdmin(admin.ModelAdmin):
     list_display = ("id", "name")
     search_fields = ("name",)
 
 
-# Admin config for Tag
 @admin.register(Tag)
 class TagAdmin(admin.ModelAdmin):
     list_display = ("id", "name")
     search_fields = ("name",)
 
 
-# Admin config for Task
 @admin.register(Task)
 class TaskAdmin(admin.ModelAdmin):
     list_display = ("id", "name", "status", "priority", "deadline", "created_by")
@@ -40,7 +23,6 @@ class TaskAdmin(admin.ModelAdmin):
     readonly_fields = ("created_time", "updated_time")
 
 
-# Admin config for Comment
 @admin.register(Comment)
 class CommentAdmin(admin.ModelAdmin):
     list_display = ("id", "task", "created_by", "created_time")
@@ -48,7 +30,6 @@ class CommentAdmin(admin.ModelAdmin):
     list_filter = ("created_time",)
 
 
-# Admin config for Project
 @admin.register(Project)
 class ProjectAdmin(admin.ModelAdmin):
     list_display = ("id", "name", "description")
