+++ manager/models.py
@@ -13,27 +13,6 @@ class Project(models.Model):
         return self.name
 
 
-class Position(models.Model):
-    name = models.CharField(max_length=100, unique=True)
-
-    def __str__(self):
-        return self.name
-
-
-class Worker(AbstractUser):
-    # Custom user model with optional position
-    position = models.ForeignKey(
-        Position,
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        related_name="workers"
-    )
-
-    def __str__(self):
-        return self.username
-
-
 class TaskType(models.Model):
     name = models.CharField(max_length=100, unique=True)
