+++ manager/tests.py
@@ -1,4 +1,3 @@
-
 from django.test import TestCase, Client
 from django.urls import reverse
 from django.contrib.auth import get_user_model
