+++ requirements.txt
@@ -1,0 +1,7 @@
+Django==5.2
+django-crispy-forms==2.4
+crispy-bootstrap5==2025.4
+django-htmx==1.23.0
+sqlparse==0.5.3
+tzdata==2025.2
+
